+++ Library/Configuration/Settings.cs
@@ -1,4 +1,4 @@
-﻿using System;
+using System;
 using System.Reflection;
 using System.Runtime.CompilerServices;
 using System.Text;
@@ -16,6 +16,7 @@ namespace Recurly.Configuration
         public int PageSize { get; private set; }
 
         protected const string RecurlyServerUri = "https://{0}.recurly.com/v2{1}";
+        public const string RecurlyApiVersion = "2.1";
 
         // static, unlikely to change
         public string UserAgent