+++ History.md
@@ -3,6 +3,9 @@ Unreleased
 
 * added; `applies_to_non_plan_charges` attribute `Coupon`
 * added; `Coupons` to `Subscription`
+* added; `uuid` to `CouponRedemption`
+* added; `GetActiveRedemptions` to `Account`
+* added; `X-Api-Version` to `2.1`
 
 1.2.2 (stable) / 2015-07-06
 ==================